backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
dj-rest-auth==2.1.2
django-admin-interface==0.14.1
django-allauth==0.44.0
django-colorfield==0.3.2
django-csvimport==2.16
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-location-field==2.1.0
django==3.1.4
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
# pyjwt==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests-oauthlib==1.3.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
uwsgi==2.0.19.1
